segment_transcript_by_topic/requirements.txt
openai-whisper
ffmpeg-python
pydub
sentence-transformers
hdbscan
numpy
tqdm 